+++ src/databricks/labs/ucx/cli.py
@@ -11,10 +11,7 @@ from databricks.labs.ucx.config import AccountConfig, ConnectConfig
 from databricks.labs.ucx.framework.crawlers import StatementExecutionBackend
 from databricks.labs.ucx.framework.tui import Prompts
 from databricks.labs.ucx.hive_metastore import ExternalLocations, TablesCrawler
-from databricks.labs.ucx.hive_metastore.mapping import (
-    ExternalLocationMapping,
-    TableMapping,
-)
+from databricks.labs.ucx.hive_metastore.mapping import TableMapping
 from databricks.labs.ucx.install import WorkspaceInstaller
 from databricks.labs.ucx.installer import InstallationManager
 
@@ -96,16 +93,15 @@ def create_table_mapping():
 def validate_external_locations():
     ws = WorkspaceClient()
     prompts = Prompts()
-    location_mapping = ExternalLocationMapping(ws)
     installation_manager = InstallationManager(ws)
     installation = installation_manager.for_user(ws.current_user.me())
     sql_backend = StatementExecutionBackend(ws, installation.config.warehouse_id)
     location_crawler = ExternalLocations(ws, sql_backend, installation.config.inventory_database)
-    path = location_mapping.save(location_crawler)
+    path = location_crawler.save()
     if len(path) > 0:
         return
     if (
-        prompts.question(f"external_locations.tf file written to {path}. Do you want to open it", default="Yes")
+        prompts.choice(f"external_locations.tf file written to {path}. Do you want to open it", choices=["Yes", "No"])
         == "Yes"
     ):
         webbrowser.open(f"{ws.config.host}/#workspace{path}")
