+++ src/databricks/labs/ucx/hive_metastore/mapping.py
@@ -11,8 +11,7 @@ from databricks.sdk.service.workspace import ImportFormat
 
 from databricks.labs.ucx.account import WorkspaceInfo
 from databricks.labs.ucx.framework.crawlers import StatementExecutionBackend
-from databricks.labs.ucx.hive_metastore import ExternalLocations, TablesCrawler
-from databricks.labs.ucx.hive_metastore.locations import ExternalLocation
+from databricks.labs.ucx.hive_metastore import TablesCrawler
 from databricks.labs.ucx.hive_metastore.tables import Table
 
 logger = logging.getLogger(__name__)
@@ -107,62 +106,3 @@ class TableMapping:
                 logger.error(nf)
         except BadRequest as br:
             logger.error(br)
-
-
-class ExternalLocationMapping:
-    def __init__(self, ws: WorkspaceClient, folder: str | None = None):
-        if not folder:
-            folder = f"/Users/{ws.current_user.me().user_name}/.ucx"
-        self._ws = ws
-        self._folder = folder
-
-    def _get_ext_location_definitions(self, missing_locations: list[ExternalLocation]) -> list:
-        tf_script = []
-        cnt = 1
-        for loc in missing_locations:
-            script = f'resource "databricks_external_location" "name_{cnt}" {{ \n'
-            script += f'name = "name_{cnt}"\n'
-            script += f'url  = "{loc.location}"\n'
-            script += "credential_name = <storage_credential_reference>\n"
-            script += "}\n"
-            tf_script.append(script)
-            cnt += 1
-        return tf_script
-
-    def _match_table_external_locations(self, locations: ExternalLocations) -> tuple[list, list[ExternalLocation]]:
-        external_locations = list(self._ws.external_locations.list())
-        location_path = [_.url for _ in external_locations]
-        table_locations = locations.snapshot()
-        matching_locations = []
-        missing_locations = []
-        for loc in table_locations:
-            if loc.location in location_path:
-                matching_locations.append([external_locations[location_path.index(loc.location)].name, loc.table_count])
-                continue
-            missing_locations.append(loc)
-        return matching_locations, missing_locations
-
-    def save(self, locations: ExternalLocations) -> str:
-        matching_locations, missing_locations = self._match_table_external_locations(locations)
-        if len(matching_locations) > 0:
-            logger.info("following external locations are already configured.")
-            logger.info("sharing details of # tables that can be migrated for each location")
-            for _ in matching_locations:
-                logger.info(f"{_[1]} tables can be migrated using external location {_[0]}.")
-        if len(missing_locations) > 0:
-            logger.info("following external location need to be created.")
-            for _ in missing_locations:
-                logger.info(f"{_.table_count} tables can be migrated using external location {_.location}.")
-            buffer = io.StringIO()
-            for script in self._get_ext_location_definitions(missing_locations):
-                buffer.write(script)
-            buffer.seek(0)
-            return self._overwrite_mapping(buffer)
-        else:
-            logger.info("no additional external location to be created.")
-            return str(None)
-
-    def _overwrite_mapping(self, buffer) -> str:
-        path = f"{self._folder}/external_locations.tf"
-        self._ws.workspace.upload(path, buffer, overwrite=True, format=ImportFormat.AUTO)
-        return path
