+++ tests/integration/hive_metastore/test_external_locations.py
@@ -1,6 +1,10 @@
 import logging
 
-from databricks.labs.ucx.hive_metastore.locations import ExternalLocations, Mount
+from databricks.labs.ucx.hive_metastore.locations import (
+    ExternalLocation,
+    ExternalLocations,
+    Mount,
+)
 from databricks.labs.ucx.hive_metastore.tables import Table
 
 logger = logging.getLogger(__name__)
@@ -81,3 +85,15 @@ def test_external_locations(ws, sql_backend, inventory_schema, env_or_skip):
     assert results[4].location == "jdbc:providerknown://somedb.us-east-1.rds.amazonaws.com:1234/test_db"
     assert results[4].table_count == 2
     assert results[5].location == "jdbc://providerunknown//somedb.us-east-1.rds.amazonaws.com:1234/test_db"
+
+
+def test_save_external_location_mapping_missing_location(ws, sql_backend, inventory_schema):
+    logger.info("setting up fixtures")
+    locations = [
+        ExternalLocation("abfss://cont1@storage123/test_location", 2),
+        ExternalLocation("abfss://cont1@storage456/test_location2", 1),
+    ]
+    sql_backend.save_table(f"{inventory_schema}.external_locations", locations, ExternalLocation)
+    location_crawler = ExternalLocations(ws, sql_backend, inventory_schema)
+    path = location_crawler.save()
+    assert ws.workspace.get_status(path)
