+++ tests/unit/hive_metastore/test_locations.py
@@ -1,6 +1,7 @@
 from unittest.mock import MagicMock, Mock
 
 from databricks.sdk.dbutils import MountInfo
+from databricks.sdk.service.catalog import ExternalLocationInfo
 
 from databricks.labs.ucx.hive_metastore.locations import (
     ExternalLocations,
@@ -108,3 +109,61 @@ def test_external_locations():
     assert result_set[4].location == "jdbc:mysql://somemysql.us-east-1.rds.amazonaws.com:3306/test_db"
     assert result_set[5].location == "jdbc:providerknown://somedb.us-east-1.rds.amazonaws.com:1234/test_db"
     assert result_set[6].location == "jdbc:providerunknown://somedb.us-east-1.rds.amazonaws.com:1234/test_db"
+
+
+def make_row(data, columns):
+    row = Row(data)
+    row.__columns__ = columns
+    return row
+
+
+def test_save_external_location_mapping_missing_location():
+    ws = MagicMock()
+    select_cols = ["location", "storage_properties"]
+    sbe = MockBackend(
+        rows={
+            "SELECT location, storage_properties FROM test.tables WHERE location IS NOT NULL": [
+                make_row(("s3://test_location/test1/table1", ""), select_cols),
+                make_row(("gcs://test_location2/test2/table2", ""), select_cols),
+                make_row(("abfss://cont1@storagetest1.dfs.core.windows.net/test2/table3", ""), select_cols),
+            ],
+        }
+    )
+    location_crawler = ExternalLocations(ws, sbe, "test", "~/.ucx")
+    ws.external_locations.list.return_value = [ExternalLocationInfo(name="loc1", url="s3://test_location/test11")]
+    location_crawler.save()
+    (path, content), _ = ws.workspace.upload.call_args
+    assert "~/.ucx/external_locations.tf" == path
+    assert (
+        'resource "databricks_external_location" "test_location_test1" { \n'
+        '    name = "test_location_test1"\n'
+        '    url  = "s3://test_location/test1"\n'
+        "    credential_name = databricks_storage_credential.<storage_credential_reference>.id\n"
+        "}\n"
+        'resource "databricks_external_location" "test_location2_test2" { \n'
+        '    name = "test_location2_test2"\n'
+        '    url  = "gcs://test_location2/test2"\n'
+        "    credential_name = databricks_storage_credential.<storage_credential_reference>.id\n"
+        "}\n"
+        'resource "databricks_external_location" "storagetest1_test2" { \n'
+        '    name = "storagetest1_test2"\n'
+        '    url  = "abfss://cont1@storagetest1.dfs.core.windows.net/test2"\n'
+        "    credential_name = databricks_storage_credential.<storage_credential_reference>.id\n"
+        "}\n"
+    ) == content.read()
+
+
+def test_save_external_location_mapping_no_missing_location():
+    ws = MagicMock()
+    select_cols = ["location", "storage_properties"]
+    sbe = MockBackend(
+        rows={
+            "SELECT location, storage_properties FROM test.tables WHERE location IS NOT NULL": [
+                make_row(("s3://test_location/test1/table1", ""), select_cols),
+            ],
+        }
+    )
+    location_crawler = ExternalLocations(ws, sbe, "test", "~/.ucx")
+    ws.external_locations.list.return_value = [ExternalLocationInfo(name="loc1", url="s3://test_location/test1")]
+    location_crawler.save()
+    ws.workspace.upload.assert_not_called()
