+++ tests/unit/hive_metastore/test_mapping.py
@@ -3,18 +3,9 @@ from unittest.mock import MagicMock, create_autospec
 
 import pytest
 from databricks.sdk.errors import NotFound
-from databricks.sdk.service.catalog import ExternalLocationInfo
 
 from databricks.labs.ucx.account import WorkspaceInfo
-from databricks.labs.ucx.hive_metastore.locations import (
-    ExternalLocation,
-    ExternalLocations,
-)
-from databricks.labs.ucx.hive_metastore.mapping import (
-    ExternalLocationMapping,
-    Rule,
-    TableMapping,
-)
+from databricks.labs.ucx.hive_metastore.mapping import Rule, TableMapping
 from databricks.labs.ucx.hive_metastore.tables import Table, TablesCrawler
 
 
@@ -140,31 +131,3 @@ def test_skip_missing_table(mocker, caplog):
     mapping = TableMapping(ws)
     mapping.skip_table(sbe, schema="schema", table="table")
     assert [rec.message for rec in caplog.records if "table not found" in rec.message.lower()]
-
-
-def test_save_external_location_mapping_missing_location():
-    ws = MagicMock()
-    ext_location_mapping = ExternalLocationMapping(ws, "~/.ucx")
-    location_crawler = create_autospec(ExternalLocations)
-    location_crawler.snapshot.return_value = [ExternalLocation(location="s3://test_location/test1", table_count=1)]
-    ws.external_locations.list.return_value = [ExternalLocationInfo(name="loc1", url="s3://test_location/test11")]
-    ext_location_mapping.save(location_crawler)
-    (path, content), _ = ws.workspace.upload.call_args
-    assert "~/.ucx/external_locations.tf" == path
-    assert (
-        'resource "databricks_external_location" "name_1" { \n'
-        'name = "name_1"\n'
-        'url  = "s3://test_location/test1"\n'
-        "credential_name = <storage_credential_reference>\n"
-        "}\n"
-    ) == content.read()
-
-
-def test_save_external_location_mapping_no_missing_location():
-    ws = MagicMock()
-    ext_location_mapping = ExternalLocationMapping(ws, "~/.ucx")
-    location_crawler = create_autospec(ExternalLocations)
-    location_crawler.snapshot.return_value = [ExternalLocation(location="s3://test_location/test1", table_count=1)]
-    ws.external_locations.list.return_value = [ExternalLocationInfo(name="loc1", url="s3://test_location/test1")]
-    ext_location_mapping.save(location_crawler)
-    ws.workspace.upload.assert_not_called()
